+++ README.md
@@ -1,1 +1,10 @@
 # fb-messenger-hackathon
+
+### Installation
+
+`npm i`
+
+### Routes 
+
+`https://<url>/api/test`
+`https://<url>/api/webhooks`
