+++ models/User.js
@@ -1,0 +1,27 @@
+const mongoose = require('mongoose');
+
+const userSchema = new mongoose.Schema({
+    id: String,
+    name: String
+});
+
+const User = mongoose.model('users',userSchema);
+
+// Create user if user not found in database
+function createUser(fbid,name){
+    var newUser = new User({
+        id:fbid,
+        name:name
+    });
+    newUser.save().then(doc => console.log(doc)).catch(err => console.log(err));
+}
+
+function checkUser(fbid){
+    return User.find({'id':fbid}).then(function(user){
+        return user;
+    }).catch(function(err){
+        console.log(err);
+    });
+}
+
+export { checkUser, createUser };
