+++ routes/api/index.js
@@ -1,0 +1,7 @@
+var router = require('express').Router();
+
+router.use('/', require('./webhooks'));
+router.use('/test', require('./test'));
+
+
+module.exports = router;
